requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
drf-spectacular>=0.26.0
drf-spectacular-sidecar
drf-spectacular